@@ -1,23 +1,22 @@
 ﻿using System;
 
-namespace NAPS2.Wia
+namespace NAPS2.Wia;
+
+public class WiaDeviceInfo : NativeWiaObject, IWiaDeviceProps
 {
-    public class WiaDeviceInfo : NativeWiaObject, IWiaDeviceProps
+    protected internal WiaDeviceInfo(WiaVersion version, IntPtr propStorageHandle) : base(version)
     {
-        protected internal WiaDeviceInfo(WiaVersion version, IntPtr propStorageHandle) : base(version)
-        {
-            Properties = new WiaPropertyCollection(version, propStorageHandle);
-        }
+        Properties = new WiaPropertyCollection(version, propStorageHandle);
+    }
 
-        public WiaPropertyCollection Properties { get; }
+    public WiaPropertyCollection Properties { get; }
 
-        protected override void Dispose(bool disposing)
+    protected override void Dispose(bool disposing)
+    {
+        base.Dispose(disposing);
+        if (disposing)
         {
-            base.Dispose(disposing);
-            if (disposing)
-            {
-                Properties?.Dispose();
-            }
+            Properties?.Dispose();
         }
     }
 }
