@@ -1,11 +1,10 @@
-﻿namespace NAPS2.Wia
+﻿namespace NAPS2.Wia;
+
+/// <summary>
+/// Property type constants.
+/// </summary>
+public static class WiaPropertyType
 {
-    /// <summary>
-    /// Property type constants.
-    /// </summary>
-    public static class WiaPropertyType
-    {
-        public const int I4 = 3;
-        public const int BSTR = 8;
-    }
+    public const int I4 = 3;
+    public const int BSTR = 8;
 }
